+++ do-release.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-cargo release --package ubi "$@"
-cargo release --package ubi-cli "$@"